in server: changed UserController to AuthController and created new UserController for user query. In the client ive moved to auth routes and moved user-data fetching to userProvider
