+++ GymnastLink-Client/src/contexts/UserContext.tsx
@@ -1,5 +1,9 @@
-import {FC, PropsWithChildren, createContext, useContext, useState} from 'react';
+import {getCurrentUserData} from '@services/api';
+import Cookies from 'js-cookie';
+import {FC, PropsWithChildren, createContext, useContext, useEffect, useState} from 'react';
+import {useNavigate} from 'react-router';
 import {LoggedUser} from '@customTypes/User';
+import {ClientRoutes} from '@enums/clientRoutes';
 
 interface UserContextType {
   user: LoggedUser | null;
@@ -10,6 +14,24 @@ const UserContext = createContext<UserContextType | undefined>(undefined);
 
 const UserProvider: FC<PropsWithChildren> = ({children}) => {
   const [user, setUser] = useState<LoggedUser | null>(null);
+  const navigate = useNavigate();
+  const token = Cookies.get('access_token');
+
+  useEffect(() => {
+    if (token) {
+      if (!user) {
+        getCurrentUserData().then((userData) => {
+          setUser(userData);
+
+          if (location.pathname === '/') {
+            navigate(ClientRoutes.UPDATES);
+          }
+        });
+      } else if (location.pathname === '/') {
+        navigate(ClientRoutes.UPDATES);
+      }
+    }
+  }, [location.pathname]);
 
   return <UserContext.Provider value={{user, setUser}}>{children}</UserContext.Provider>;
 };
