+++ GymnastLink-Client/src/enums/serverRoutes.ts
@@ -1,0 +1,8 @@
+enum ServerRoutes {
+  USER = 'users',
+  POSTS = 'posts',
+  COMMENTS = 'comments',
+  AUTH = 'auth',
+}
+
+export {ServerRoutes};
