+++ GymnastLink-Client/src/services/api.ts
@@ -3,19 +3,20 @@ import axios from 'axios';
 import Cookies from 'js-cookie';
 import {UserLoginForm} from '@components/LoginForm/form';
 import {LoggedUser} from '@customTypes/User';
+import {ServerRoutes} from '@enums/serverRoutes';
 import {parseExpirationInDays} from '@utils/dateUtils';
 
 const BASE_URL = import.meta.env.VITE_SERVER_URL;
 const JWT_TOKEN_EXPIRES = import.meta.env.VITE_JWT_TOKEN_EXPIRES;
 
 const registerUser = async (loginForm: UserLoginForm) => {
-  const response = await axios.post<LoggedUser>(`${BASE_URL}/users/register`, loginForm);
+  const response = await axios.post<LoggedUser>(`${BASE_URL}/${ServerRoutes.AUTH}/register`, loginForm);
 
   return response.data;
 };
 
 const userLogin = async (loginForm: UserLoginForm) => {
-  const response = await axios.post<LoggedUser>(`${BASE_URL}/users/login`, loginForm);
+  const response = await axios.post<LoggedUser>(`${BASE_URL}/${ServerRoutes.AUTH}/login`, loginForm);
   Cookies.set('access_token', response.data.accessToken, {expires: parseExpirationInDays(JWT_TOKEN_EXPIRES)});
   Cookies.set('refresh_token', response.data.refreshToken);
 
@@ -26,7 +27,7 @@ const userLogout = async () => {
   const refreshToken = Cookies.get('refresh_token');
 
   if (refreshToken) {
-    await axiosInstance.post('/users/logout', {refreshToken});
+    await axiosInstance.post(`/${ServerRoutes.AUTH}/logout`, {refreshToken});
 
     Cookies.remove('access_token');
     Cookies.remove('refresh_token');
@@ -35,10 +36,10 @@ const userLogout = async () => {
   }
 };
 
-const getUserData = async () => {
-  const response = await axiosInstance.get<LoggedUser>('/users/user-data');
+const getCurrentUserData = async () => {
+  const response = await axiosInstance.get<LoggedUser>(`/${ServerRoutes.AUTH}/user-data`);
 
   return response.data;
 };
 
-export {registerUser, userLogin, userLogout, getUserData};
+export {registerUser, userLogin, userLogout, getCurrentUserData};
