+++ GymnastLink-Server/src/models/usersModel.ts
@@ -5,6 +5,7 @@ interface IUser {
   email: string;
   password: string;
   userName: string;
+  profileImage?: string;
   refreshToken?: string[];
 }
 
@@ -22,6 +23,7 @@ const userSchema = new Schema<IUser>({
     type: String,
     required: true,
   },
+  profileImage: String,
   refreshToken: {
     type: [String],
     default: [],
