+++ GymnastLink-Server/src/routes/commentsRoutes.ts
@@ -1,6 +1,6 @@
 import {Router} from 'express';
 import commentsController from '../controllers/commentsController';
-import {authMiddleware} from '../controllers/usersController';
+import {authMiddleware} from '../controllers/authController';
 
 const router = Router();
 
@@ -55,7 +55,11 @@ const router = Router();
  *       500:
  *         description: Server error
  */
-router.get('/', commentsController.getAll.bind(commentsController));
+router.get(
+  '/',
+  authMiddleware,
+  commentsController.getAll.bind(commentsController)
+);
 
 /**
  * @swagger
@@ -83,7 +87,11 @@ router.get('/', commentsController.getAll.bind(commentsController));
  *       500:
  *         description: Server error
  */
-router.get('/:id', commentsController.getById.bind(commentsController));
+router.get(
+  '/:id',
+  authMiddleware,
+  commentsController.getById.bind(commentsController)
+);
 
 /**
  * @swagger
