+++ GymnastLink-Server/src/routes/usersRoutes.ts
@@ -1,12 +1,6 @@
 import {Router} from 'express';
-import {
-  authMiddleware,
-  getUserData,
-  login,
-  logout,
-  refreshUserToken,
-  register,
-} from '../controllers/usersController';
+import usersController from '../controllers/usersController';
+import {authMiddleware} from '../controllers/authController';
 
 const router = Router();
 
@@ -19,195 +13,42 @@ const router = Router();
 
 /**
  * @swagger
- * components:
- *   securitySchemes:
- *     bearerUser:
- *       type: http
- *       scheme: bearer
- *       bearerFormat: JWT
- */
-
-/**
- * @swagger
- * components:
- *   schemas:
- *     User:
- *       type: object
- *       required:
- *         - email
- *         - password
- *       properties:
- *         email:
- *           type: string
- *           description: The user email
- *         password:
- *           type: string
- *           description: The user password
- *       example:
- *         email: 'bob@gmail.com'
- *         password: '123456'
- */
-
-/**
- * @swagger
- * components:
- *   schemas:
- *     Tokens:
- *       type: object
- *       required:
- *         - accessToken
- *         - refreshToken
- *       properties:
- *         accessToken:
- *           type: string
- *           description: The JWT access token
- *         refreshToken:
- *           type: string
- *           description: The JWT refresh token
- *       example:
- *         accessToken: '123cd123x1xx1'
- *         refreshToken: '134r2134cr1x3c'
- */
-
-/**
- * @swagger
- * /users/register:
- *   post:
- *     summary: registers a new user
+ * /users/{id}:
+ *   get:
+ *     summary: Get user by ID
  *     tags: [User]
- *     requestBody:
- *       required: true
- *       content:
- *         application/json:
- *           schema:
- *             $ref: '#/components/schemas/User'
+ *     parameters:
+ *       - in: path
+ *         name: id
+ *         schema:
+ *           type: string
+ *         required: true
+ *         description: The user ID
  *     responses:
  *       200:
- *         description: The new user
- *         content:
- *           application/json:
- *             schema:
- *               $ref: '#/components/schemas/User'
- */
-router.post('/register', register);
-
-/**
- * @swagger
- * /users/login:
- *   post:
- *     summary: User login
- *     description: Authenticate user and return tokens
- *     tags: [User]
- *     requestBody:
- *       required: true
- *       content:
- *         application/json:
- *           schema:
- *             $ref: '#/components/schemas/User'
- *     responses:
- *       200:
- *         description: Successful login
- *         content:
- *           application/json:
- *             schema:
- *               $ref: '#/components/schemas/Tokens'
- *       400:
- *         description: Invalid credentials or request
- *       500:
- *         description: Server error
- */
-router.post('/login', login);
-
-/**
- * @swagger
- * /users/logout:
- *   post:
- *     summary: User logout
- *     description: Logout user and invalidate the refresh token
- *     tags: [User]
- *     security:
- *       - bearerAuth: []
- *     requestBody:
- *       required: true
- *       content:
- *         application/json:
- *           schema:
- *             type: object
- *             properties:
- *               refreshToken:
- *                 type: string
- *                 example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
- *     responses:
- *       200:
- *         description: Successful logout
- *       400:
- *         description: Invalid refresh token
- *       500:
- *         description: Server error
- */
-router.post('/logout', logout);
-
-/**
- * @swagger
- * /users/refresh-token:
- *   post:
- *     summary: Refresh tokens
- *     description: Refresh access and refresh tokens using the provided refresh token
- *     tags: [User]
- *     security:
- *       - bearerAuth: []
- *     requestBody:
- *       required: true
- *       content:
- *         application/json:
- *           schema:
- *             type: object
- *             properties:
- *               refreshToken:
- *                 type: string
- *                 example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
- *     responses:
- *       200:
- *         description: Tokens refreshed successfully
+ *         description: The user data
  *         content:
  *           application/json:
  *             schema:
  *               type: object
  *               properties:
- *                 accessToken:
+ *                 id:
  *                   type: string
- *                   example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
- *                 refreshToken:
+ *                   description: The user ID
+ *                   example: 60d0fe4f5311236168a109ca
+ *                 userName:
  *                   type: string
- *                   example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
+ *                   description: The user's name
+ *                   example: John Doe
+ *                 profileImage:
+ *                   type: string
+ *                   description: The user's profile image URL
+ *                   example: http://example.com/image.jpg
+ *       404:
+ *         description: User not found
  *       400:
- *         description: Invalid refresh token
- *       500:
- *         description: Server error
+ *         description: Bad request
  */
-router.post('/refresh-token', refreshUserToken);
-
-/**
- * @swagger
- * /users/user-data:
- *   get:
- *     summary: Get user data
- *     description: Retrieve user data using the access token
- *     tags: [User]
- *     security:
- *       - bearerAuth: []
- *     responses:
- *       200:
- *         description: User data
- *         content:
- *           application/json:
- *             schema:
- *               $ref: '#/components/schemas/User'
- *       401:
- *         description: Unauthorized
- *       500:
- *         description: Server error
- */
-router.get('/user-data', authMiddleware, getUserData);
+router.get('/:id', authMiddleware, usersController.getUserById);
 
 export {router as userRouter};
