+++ GymnastLink-Server/src/types/UserResponse.ts
@@ -1,0 +1,7 @@
+interface UserResponse {
+  id: string;
+  userName: string;
+  profileImage?: string;
+}
+
+export type {UserResponse};
